+++ src/main/java/org/grnet/pidmr/service/DatabaseProviderService.java
@@ -9,9 +9,11 @@ import jakarta.transaction.Transactional;
 import jakarta.ws.rs.NotAcceptableException;
 import jakarta.ws.rs.NotFoundException;
 import jakarta.ws.rs.core.UriInfo;
+import org.apache.commons.lang3.ArrayUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.grnet.pidmr.dto.*;
 import org.grnet.pidmr.entity.database.Action;
+import org.grnet.pidmr.entity.database.Endpoint;
 import org.grnet.pidmr.entity.database.Provider;
 import org.grnet.pidmr.entity.database.Regex;
 import org.grnet.pidmr.enums.MailType;
@@ -129,7 +131,7 @@ public class DatabaseProviderService implements ProviderServiceI {
             var identification = new Identification();
             identification.status = Identification.Status.INVALID;
             identification.type = "";
-            identification.example = "";
+            identification.examples = new String[]{""};
             identifications.add(identification);
         } else {
 
@@ -148,7 +150,7 @@ public class DatabaseProviderService implements ProviderServiceI {
 
     private void check(CharSequence cs, Pattern pattern, Provider provider, Set<Identification> identifications) {
 
-        Matcher matcher = pattern.matcher(cs);
+        var matcher = pattern.matcher(cs);
 
         var dto = ProviderMapper.INSTANCE.databaseProviderToDto(provider);
 
@@ -159,7 +161,7 @@ public class DatabaseProviderService implements ProviderServiceI {
                 var identification = new Identification();
                 identification.status = Identification.Status.VALID;
                 identification.type = provider.getType();
-                identification.example = provider.getExample();
+                identification.examples = provider.getExamples();
                 identification.actions = dto.actions;
                 identifications.add(identification);
             } else {
@@ -172,7 +174,7 @@ public class DatabaseProviderService implements ProviderServiceI {
             var identification = new Identification();
             identification.status = Identification.Status.AMBIGUOUS;
             identification.type = provider.getType();
-            identification.example = provider.getExample();
+            identification.examples = provider.getExamples();
             identification.actions = dto.actions;
             identifications.add(identification);
         }
@@ -180,7 +182,7 @@ public class DatabaseProviderService implements ProviderServiceI {
 
     private Identification check(CharSequence cs, Pattern pattern, Provider provider, Identification identification) {
 
-        Matcher matcher = pattern.matcher(cs);
+        var matcher = pattern.matcher(cs);
 
         var dto = ProviderMapper.INSTANCE.databaseProviderToDto(provider);
 
@@ -190,7 +192,7 @@ public class DatabaseProviderService implements ProviderServiceI {
 
                 identification.status = Identification.Status.VALID;
                 identification.type = provider.getType();
-                identification.example = provider.getExample();
+                identification.examples = provider.getExamples();
                 identification.actions = dto.actions;
             }
             return identification;
@@ -200,7 +202,7 @@ public class DatabaseProviderService implements ProviderServiceI {
 
             identification.status = Identification.Status.AMBIGUOUS;
             identification.type = provider.getType();
-            identification.example = provider.getExample();
+            identification.examples = provider.getExamples();
             identification.actions = dto.actions;
         }
 
@@ -215,7 +217,7 @@ public class DatabaseProviderService implements ProviderServiceI {
         newProvider.setName(request.name);
         newProvider.setType(request.type);
         newProvider.setDescription(request.description);
-        newProvider.setExample(request.example);
+        newProvider.setExamples(request.examples);
         newProvider.setCreatedBy(requestUserContext.getVopersonID());
         newProvider.setStatus(ProviderStatus.PENDING);
         newProvider.setReliesOnDois(request.reliesOnDois);
@@ -268,8 +270,9 @@ public class DatabaseProviderService implements ProviderServiceI {
      */
     @Transactional
     public ProviderDto createV2(ProviderRequestV2 request) {
-        SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
-        // Set the formatter's time zone to UTC
+
+        var formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
+
         formatter.setTimeZone(TimeZone.getTimeZone("UTC"));
 
         checkIfActionsSupported(request.actions.stream().map(action -> action.mode).collect(Collectors.toSet()));
@@ -281,9 +284,11 @@ public class DatabaseProviderService implements ProviderServiceI {
         request
                 .actions
                 .forEach
-                        (action -> {
-                            newProvider.addAction(actionRepository.findById(action.mode), action.endpoints);
-                        });
+                        (action -> newProvider.addAction(actionRepository.findById(action.mode), Arrays.stream(action
+                                .endpoints)
+                                .map(endpoint -> new Endpoint(endpoint, request.type))
+                                .collect(Collectors.toList())
+                        ));
 
         request.
                 regexes
@@ -306,6 +311,51 @@ public class DatabaseProviderService implements ProviderServiceI {
         return ProviderMapper.INSTANCE.databaseProviderToDto(newProvider);
     }
 
+    /**
+     * This method stores a new Provider in the database.
+     *
+     * @param request The Provider to be created.
+     * @return The created Provider.
+     * @throws
+     */
+    @Transactional
+    public ProviderDto createV3(ProviderRequestV3 request) {
+
+        var formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
+
+        formatter.setTimeZone(TimeZone.getTimeZone("UTC"));
+
+        checkIfActionsSupported(request.actions.stream().map(action -> action.mode).collect(Collectors.toSet()));
+
+        var newProvider = setProviderForCreation(request);
+
+        newProvider.setValidator(Validator.valueOf(request.validator));
+
+        request
+                .actions
+                .forEach
+                        (action -> newProvider.addAction(actionRepository.findById(action.mode), action.endpoints));
+
+        request.
+                regexes
+                .forEach(regex -> {
+                    var regexp = new Regex();
+                    regexp.setRegex(regex);
+                    newProvider.addRegex(regexp);
+                });
+
+
+        providerRepository.persist(newProvider);
+        var timestamp = formatter.format(Timestamp.from(Instant.now()));
+
+        var userID = newProvider.getCreatedBy();
+        var pidtype = newProvider.getType();
+
+        var emailContext = new EmailContextForStatusUpdate(userID, keycloakAdminService.getUserEmail(userID), newProvider.getId(), String.valueOf(newProvider.getStatus()), pidtype,timestamp);
+        mailerService.sendEmailsWithContext(emailContext, MailType.ADMIN_ALERT_NEW_PID_TYPE_ENTRY_CREATION, MailType.PROVIDER_ADMIN_NEW_PID_TYPE_ENTRY_CREATION);
+
+        return ProviderMapper.INSTANCE.databaseProviderToDto(newProvider);
+    }
 
     /**
      * This method deletes from database a Provider by its ID.
@@ -366,8 +416,8 @@ public class DatabaseProviderService implements ProviderServiceI {
         if (StringUtils.isNotEmpty(request.description)) {
             provider.setDescription(request.description);
         }
-        if (StringUtils.isNotEmpty(request.example)) {
-            provider.setExample(request.example);
+        if (ArrayUtils.isNotEmpty(request.examples)) {
+            provider.setExamples(request.examples);
         }
         provider.setReliesOnDois(request.reliesOnDois);
         provider.setStatus(ProviderStatus.PENDING);
@@ -428,8 +478,41 @@ public class DatabaseProviderService implements ProviderServiceI {
             var actions = provider.getActions();
             new ArrayList<>(actions).forEach(action -> provider.removeAction(action.getAction()));
             Panache.getEntityManager().flush();
+            request.actions.forEach(newAction -> provider.addAction(actionRepository.findById(newAction.mode), Arrays.stream(newAction
+                            .endpoints)
+                    .map(endpoint -> new Endpoint(endpoint, provider.getType()))
+                    .collect(Collectors.toList())));
+        }
+
+        return ProviderMapper.INSTANCE.databaseProviderToDto(provider);
+    }
+
+    /**
+     * This method updates one or more attributes of a Provider.
+     *
+     * @param request The Provider attributes to be updated.
+     * @param id      The Provider to be updated.
+     * @return The updated Provider.
+     */
+    @ManageEntity(entityType = "Provider")
+    @Transactional
+    public ProviderDto updateV3(Long id, UpdateProviderV3 request) {
+
+        var provider = setProviderForUpdating(id, request);
+
+        if (StringUtils.isNotEmpty(request.validator)) {
+            provider.setValidator(Validator.valueOf(request.validator));
+        }
+
+        if (!request.actions.isEmpty()) {
+
+            checkIfActionsSupported(request.actions.stream().map(action -> action.mode).collect(Collectors.toSet()));
+            var actions = provider.getActions();
+            new ArrayList<>(actions).forEach(action -> provider.removeAction(action.getAction()));
+            Panache.getEntityManager().flush();
             request.actions.forEach(newAction -> provider.addAction(actionRepository.findById(newAction.mode), newAction.endpoints));
         }
+
         return ProviderMapper.INSTANCE.databaseProviderToDto(provider);
     }
 
@@ -489,20 +572,14 @@ public class DatabaseProviderService implements ProviderServiceI {
      */
     @Transactional
     public AdminProviderDto updateProviderStatus(Long id, ProviderStatus status) {
-        SimpleDateFormat formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
-        // Set the formatter's time zone to UTC
+
+        var formatter = new SimpleDateFormat("dd-MM-yyyy HH:mm:ss");
+
         formatter.setTimeZone(TimeZone.getTimeZone("UTC"));
 
         var newProvider = providerRepository.findById(id);
         newProvider.setStatus(status);
         newProvider.setStatusUpdatedBy(requestUserContext.getVopersonID());
-
-//        var pidtype=newProvider.getType();
-//        var userID = newProvider.getCreatedBy();
-//        var timestamp = formatter.format(Timestamp.from(Instant.now()));
-//        var emailContext = new EmailContextForStatusUpdate(userID, keycloakAdminService.getUserEmail(userID), newProvider.getId(), String.valueOf(status), pidtype,timestamp);
-//        mailerService.sendEmailsWithContext(emailContext, MailType.PROVIDER_ADMIN_ALERT_CHANGE_PID_TYPE_ENTRY_REQUEST_STATUS, MailType.PROVIDER_ADMIN_ALERT_CHANGE_PID_TYPE_ENTRY_REQUEST_STATUS);
-
 
         return ProviderMapper.INSTANCE.databaseAdminProviderToDto(newProvider);
     }
@@ -515,7 +592,7 @@ public class DatabaseProviderService implements ProviderServiceI {
         var identification = new Identification();
         identification.status = Identification.Status.INVALID;
         identification.type = "";
-        identification.example = "";
+        identification.examples = new String[]{""};
 
         for (Regex regex : regexes) {
 
@@ -529,5 +606,4 @@ public class DatabaseProviderService implements ProviderServiceI {
 
         return identification;
     }
-
 }